# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Reserva de Laboratórios e Salas",
    "version": "1.0.0",
    "description": "Importe no Insomnia. Cadastre usuário, confira login e configure Basic Auth. Use IDs retornados nas consultas. Datas DD/MM/AAAA e horários HH:MM:SS. Testes ainda pendentes."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "security": [
    {
      "basicAuth": []
    }
  ],
  "paths": {
...
    "/api/status-recursos": {
      "get": {
        "summary": "Consultar status-recursos",
        "tags": [
          "status-recursos"
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StatusView"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cadastrar status-recursos",
        "tags": [
          "status-recursos"
        ],
        "responses": {
          "201": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusView"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "URI do registro criado",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusInput"
              }
            }
          }
        }
      }
    },
    "/api/status-recursos/{codigo}": {
      "get": {
        "summary": "Buscar por código",
        "tags": [
          "status-recursos"
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusView"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ]
      },
      "put": {
        "summary": "Atualizar cadastro completo",
        "tags": [
          "status-recursos"
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusView"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusInput"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ]
      },
      "delete": {
        "summary": "Excluir registro (reserva ativa respeita prazo de cancelamento)",
        "tags": [
          "status-recursos"
        ],
        "responses": {
          "204": {
            "description": "Sucesso"
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ]
      }
    },
...
    "/api/recursos/{codigo}": {
      "get": {
        "summary": "Buscar recurso",
        "tags": [
          "recursos"
        ],
        "responses": {
          "200": {
            "description": "Sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecursoView"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Registro não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
...
      "StatusInput": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 3,
            "maxLength": 20,
            "example": "MANUTENCAO"
          }
        },
        "required": [
          "nome"
        ],
        "additionalProperties": false
      },
      "StatusView": {
        "type": "object",
        "properties": {
          "codigo": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "nome": {
            "type": "string"
          }
        },
        "required": [
          "codigo",
          "nome"
        ],
        "additionalProperties": false
      },
...
      "RecursoView": {
        "type": "object",
        "properties": {
          "codigo": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "tipo": {
            "type": "string",
            "enum": [
              "LABORATORIO",
              "SALA"
            ]
          },
          "nome": {
            "type": "string"
          },
          "capacidade": {
            "type": "integer"
          },
          "localizacao": {
            "type": "string"
          },
          "statusRecurso": {
            "$ref": "#/components/schemas/StatusView"
          }
        },
        "required": [
          "codigo",
          "tipo",
          "nome",
          "capacidade",
          "localizacao",
          "statusRecurso"
        ],
        "additionalProperties": false
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer"
          },
          "mensagem": {
            "type": "string"
          },
          "caminho": {
            "type": "string"
          },
          "erros": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "campo": {
                  "type": "string"
                },
                "mensagem": {
                  "type": "string"
                }
              },
              "required": [
                "campo",
                "mensagem"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "timestamp",
          "status",
          "mensagem",
          "caminho",
          "erros"
        ],
        "additionalProperties": false
      }
    }
  }
}